refactor: schemas 관심사 분리
