+++ app/schemas/common_schemas.py
@@ -1,0 +1,20 @@
+from typing import Generic, Optional, List, TypeVar
+from pydantic import BaseModel
+
+T = TypeVar("T")
+
+#공통 응답 모델
+class ResponseModel(BaseModel, Generic[T]):
+    success: bool
+    data: Optional[T] = None
+   
+# 페이징 모델
+class PaginationInfo(BaseModel):
+    current_page: int
+    total_pages: int
+    total_items: int
+    items_per_page: int
+class PaginatedResponseModel(BaseModel, Generic[T]):
+    success: bool
+    data: List[T]
+    pagination: PaginationInfo
